Add very rudimentary cgc utility
The cgc utility is a wrapper around libcgdrv.so, the L4T shader compiler
provided by NVIDIA. It simply allows a file to be read and passed to the
shader compiler and outputs the resulting binary code.

The cgc utility needs to link against libcgdrv.so, so add a configure
check to enable building cgc only if the shared library exists.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>