Add host1x library
The goal of this library is to abstract away the differences between the
L4T and upstream kernel interfaces. It provides a low-level API that can
be used to build and submit command streams without worrying about which
kernel the programs run on.

Eventually everything should be migrated to use this library in order to
remove the various duplicates.

Signed-off-by: Thierry Reding <thierry@gilfi.de>