host1x: Implement VBLANK sync'ing
When displaying a new framebuffer, allow the operation to optionally
sync to VBLANK events in order to avoid tearing. For fullscreen display
this can easily be done using drmModePageFlip(), but for overlays there
isn't a good way to achieve the same. The best we can currently do is
call drmWaitVBlank() and flip as soon as possible after that. Usually it
is already too late and some tearing can't be avoided.

Signed-off-by: Thierry Reding <thierry@gilfi.de>