grate: Implement grate_draw_elements() and more
This is a bit of a code dump because it contains a lot of work done over
the last few weeks squashed into a single commit. Trying to separate
commits cleanly turned out to be too much work and error prone, so I
gave in and just committed everything in one go.

In a nutshell, this provides an implementation of grate_draw_elements()
which should be able to draw arbitrary geometry using shaders produced
by the compiler in libcgdrv. It has basic functionality to obtain the
location of attributes and uniforms so they can be properly hooked up
and allows geometry to be specified in a way similar to how vertex
buffer objects are handled by OpenGL.

Furthermore some code is added to libgrate that allows a scene to be
presented on-screen with double-buffering and synchronizing to VBLANK
events.

Signed-off-by: Thierry Reding <thierry@gilfi.de>