grate: stencil test support
For some unknown reason there is a stencil buffer distortion on a stencil write
if a linear layout is used for the stencil buffer, while tiled works fine.
The grate-3d stencil setup takes into account the fragment discard, so both
early and late stencil buffer writes are performed as expected.