libhost1x: minimize 3d registers initialization
There are a lot of redundant stuff in the 3d initialization routine, let's
get rid of it. Note that vertex/linker/fragment program cleanup with NOP's
has been removed from the 3d initialization routine because I think it is
not needed since a proper program that doesn't rely on a initial state would
be loaded anyway.