tools/assembler: dump raw data of compiled streams
Compiled commands streams would be printed to the std output. Carrying the
whole asm parser from grate to opentegra isn't an option, we may use grate's
asm tool to generate raw arrays of commands that could be cut-and-pasted to
opentegra for a use. And of course that feature could be helpful with the
debugging of the asm-parser.