add meson build system
Meson is quite a bit faster than Autotools, which seems like a good
reason to try this out:

Full rebuild, autotools:
> real 300.18
> user 235.32
> sys 32.37

Full rebuild, meson:
> real 46.92
> user 70.71
> sys 11.24

The time above is the time './configure && make', and
'meson . build && ninja -C build' took on my Toshiba AC100, measured
using '/usr/bin/time -p'.

So, less than a sixth of the time for a full rebuild, which is very
neat. However, it's on partial rebuilds that it really shines:

Partial rebuild, autotools:
> real 30.60
> user 21.07
> sys 4.27

Partial rebuild, meson:
> real 2.27
> user 1.92
> sys 0.27

In this case, I touched src/libhost1x/host1x-gr3d.c, and rebuilt
with 'make' and 'ninja' respectively. And the result is pretty
awesome, dropping to below a tenth of the build-time.