libhost1x: support DRM FB modifiers
FB modifiers are supported since 4.17 kernel and now a modifier is required
in order to display tiled framebuffer correctly. This fixes displaying of
grate tests with 4.17 kernel. Note that kernel require all parameters of the
DRM_IOCTL_MODE_ADDFB2 to be initialized when modifiers are used, hence all
offsets and pitches are now initialized for the unused planes.