tools: factor out GR3D resetting into a separate tool
It's not very difficult to get into a trouble if 3d is reset while hardware
is doing something. In a case of grate-kernel there is a good chance that
kernel driver will recover hang hardware (although not 100%, MC may hang
solidly), while vanilla upstream and downstream are likely to become unusable.
Hence, let's factor out the HW resetting feature into a separate tool.